requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# GitHub API
PyGithub==2.1.1

# HTTP Requests
requests==2.31.0

# LLM APIs (Optional - only install if not using AI Pipeline)
# anthropic==0.34.0
# openai==1.3.7

# Environment Variables
python-dotenv==1.0.0

# Additional dependencies
gitpython==3.1.40
